requirements.txt
--index-url=http://static.mediadrop.net/dependencies/dev/
iso8601
ddt
formencode >= 1.2.4 # (version required by Pylons 1.0)
Pylons >= 1.0
# WebOb 1.2.x raises an error if we use "request.str_params" (as we did in
# MediaDrop 0.10/WebOb 1.0.7) but the non-deprecated attribute was only
# added in WebOb 1.1 so we need that as baseline.
WebOb >= 1.1
WebHelpers == 1.0
# 0.7: event listener infrastructure, alembic 0.5 requires at least 0.7.3
# we need to change our class_mappers for 0.8 support
SQLAlchemy >= 0.7.3, <= 0.7.10
# theoretically every alembic since 0.4 should work (which added the
# alembic.config.Config class) but MediaDrop is only tested with 0.5+
alembic >= 0.4, <= 0.6.7
Genshi >= 0.6 # i18n improvements in Genshi
# custom translations per instance (missing Translations._catalog attribute)
Babel >= 1.0
Routes == 1.12.3
repoze.who == 1.0.18
repoze.who-friendlyform
repoze.who.plugins.sa
Paste >= 1.7.5.1 # (version required by Pylons 1.0)
PasteScript >= 1.7.4.2 # paster_plugins=
PasteDeploy >= 1.5 # (version required by Pylons 1.0)
ToscaWidgets >= 0.9.12 # 0.9.9 is not compatible with Pylons 1.0
tw.forms == 0.9.9
MySQL-python >= 1.2.2
Pillow
akismet == 0.2.0
google-api-python-client
unidecode
decorator >= 3.3.2 # (version required by Pylons 1.0)
simplejson >= 2.2.1 # (version required by Pylons 1.0)
--index-url=http://pypi.python.org/simple
aniso8601==1.0.0
bleach >= 1.4.1
python-swiftclient == 2.6.0
configobj
